Develop

+++ Controllers/AccountsController.cs
@@ -1,11 +1,15 @@
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Authorization;
+using System.Security.Claims;
 
 using Backend.DTOs;
 using Backend.Models;
 using Backend.Services;
+using Backend.Utils;
 
 namespace Backend.Controllers;
 
+[Authorize]
 [Route("api/[controller]")]
 public class AccountsController : ControllerBase
 {
@@ -37,13 +41,13 @@ public class AccountsController : ControllerBase
         return existingAccounts;
     }
 
-    [HttpPost] // POST api/Accounts
+    [HttpPost, AllowAnonymous] // POST api/Accounts
     public ActionResult Create([FromBody] AccountCreateBind body)
     {
         var newAccount = new Account
         {
             Username = body.Username,
-            Password = body.Password,
+            Password = HashHelper.Hash(body.Password),
             FirstName = body.FirstName,
             LastName = body.LastName,
             BirthDate = body.BirthDate,
@@ -54,6 +58,7 @@ public class AccountsController : ControllerBase
         return CreatedAtAction(nameof(GetDTOById), new { id = newAccount.Id }, newAccount);
     }
 
+
     [HttpPut("{id}")] // PUT api/Accounts/{id}
     public ActionResult Update(string id, [FromBody] AccountUpdateBind body)
     {
@@ -62,6 +67,11 @@ public class AccountsController : ControllerBase
         if (existingAccount == null)
         {
             return NotFound(new { message = $"Account Id:{id} is not found" });
+        }
+
+        if (ClaimHelper.GetClaim(User.Identity, ClaimTypes.Sid).Value != existingAccount.Id)
+        {
+            return Forbid();
         }
 
         existingAccount.FirstName = body.FirstName;
@@ -83,12 +93,17 @@ public class AccountsController : ControllerBase
             return NotFound(new { message = $"Account Id:{id} is not found" });
         }
 
+        if (ClaimHelper.GetClaim(User.Identity, ClaimTypes.Role).Value == "member" && ClaimHelper.GetClaim(User.Identity, ClaimTypes.Sid).Value != existingAccount.Id)
+        {
+            return Forbid();
+        }
+
         this._accountService.Remove(id);
 
         return Ok($"Account Id:{id} deleted");
     }
 
-    [HttpPut("ban/{id}")] // PUT api/Accounts/ban/{id}
+    [HttpPut("ban/{id}"), Authorize(Roles = "admin")] // PUT api/Accounts/ban/{id}
     public ActionResult Ban(string id)
     {
         var existingAccount = this._accountService.GetById(id);
@@ -105,7 +120,7 @@ public class AccountsController : ControllerBase
         return NoContent();
     }
 
-    [HttpPut("unban/{id}")] // PUT api/Accounts/unban/{id}
+    [HttpPut("unban/{id}"), Authorize(Roles = "admin")] // PUT api/Accounts/unban/{id}
     public ActionResult Unban(string id)
     {
         var existingAccount = this._accountService.GetById(id);
